pyproject.toml
[project]
name = "pyarchery"
version = "0.1.4"
description = "Python binding to Archery"
readme = "README.md"
authors = [
    { name = "Romuald Rousseau", email = "romuald.rousseau@servier.com" },
]
requires-python = "~=3.11"
dependencies = [
    "jpype1>=1.6.0",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "requests>=2.32.5",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = ["pytest>=9.0.1"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# [tool.hatch.build.targets.sdist]
# include = ["src/pyarchery", "libs", "dependencies"]

# [tool.hatch.build.targets.wheel]
# include = ["src/pyarchery", "libs", "dependencies"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "N",  # PEP8 naming convetions
    "D",  # pydocstyle
]
ignore = [
    "C901", # Too complex
    "W191", # Indentation contains tabs
    "D401", # Imperative mood
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D102", # Missing docstring in public method
    "D103", # Missing docstring in public function
    "D104", # Missing docstring in public package
    "D105", # Missing docstring in magic method
    "D106", # Missing docstring in public nested class
    "D107", # Missing docstring in `__init__`
    "D203", # Incorrect Blank Line Before Class
    "D213", # Multi-line Summary Second Line
    "E731", # Lambda-assignment
]

[tool.pytest.ini_options]
testpaths = ["tests"]
